Address review issues and especially comment the need for a dynamic_cast in a test better.
